Initial commit v2 Alexa Skills Kit SDK demo node.js
Prerequisites

An Amazon Developer account. This is required to create and configure Alexa skills.
An Amazon Web Services (AWS) account.
A NPM project with a dependency on the SDK installed as described in the Setting up the ASK SDK v2 for Node.js section. This sample skill requires either the standard SDK distribution, or if customizing your dependencies, requires that you install the support modules for the core SDK.

Installed the ASK SDK v2 for Node.js

Installed add-on package PersistenceAdapter